third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-invalid-relative-color-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['color'] = "color(from color(srgb 0.7 0.5 0.3) srgb x y z)" should not set the property value
  assert_equals: expected "" but got "color(srgb 0.7 0.5 0.3)"
[FAIL] e.style['color'] = "color(from color(srgb-linear 0.7 0.5 0.3) srgb-linear x y z)" should not set the property value
  assert_equals: expected "" but got "color(srgb-linear 0.7 0.5 0.3)"
[FAIL] e.style['color'] = "color(from color(a98-rgb 0.7 0.5 0.3) a98-rgb x y z)" should not set the property value
  assert_equals: expected "" but got "color(a98-rgb 0.7 0.5 0.3)"
[FAIL] e.style['color'] = "color(from color(rec2020 0.7 0.5 0.3) rec2020 x y z)" should not set the property value
  assert_equals: expected "" but got "color(rec2020 0.7 0.5 0.3)"
[FAIL] e.style['color'] = "color(from color(prophoto-rgb 0.7 0.5 0.3) prophoto-rgb x y z)" should not set the property value
  assert_equals: expected "" but got "color(prophoto-rgb 0.7 0.5 0.3)"
[FAIL] e.style['color'] = "color(from color(display-p3 0.7 0.5 0.3) display-p3 x y z)" should not set the property value
  assert_equals: expected "" but got "color(display-p3 0.7 0.5 0.3)"
[FAIL] e.style['color'] = "color(from color(xyz 7 -20.5 100) xyz r g b)" should not set the property value
  assert_equals: expected "" but got "color(xyz-d65 7 -20.5 100)"
[FAIL] e.style['color'] = "color(from color(xyz-d50 7 -20.5 100) xyz-d50 r g b)" should not set the property value
  assert_equals: expected "" but got "color(xyz-d50 7 -20.5 100)"
[FAIL] e.style['color'] = "color(from color(xyz-d65 7 -20.5 100) xyz-d65 r g b)" should not set the property value
  assert_equals: expected "" but got "color(xyz-d65 7 -20.5 100)"
Harness: the test ran to completion.
